abstract out Schnorr stuff
